Release a master

See merge request proyecto3/gpsapp!61

+++ src/main/webapp/app/entities/cliente/cliente-delete-dialog.component.ts
@@ -6,29 +6,68 @@ import { JhiEventManager } from 'ng-jhipster';
 
 import { ICliente } from 'app/shared/model/cliente.model';
 import { ClienteService } from './cliente.service';
+import {TareaService} from "app/entities/tarea";
+import {SubTarea} from "app/shared/model/sub-tarea.model";
+import {Tarea} from "app/shared/model/tarea.model";
 
 @Component({
     selector: 'jhi-cliente-delete-dialog',
     templateUrl: './cliente-delete-dialog.component.html'
 })
-export class ClienteDeleteDialogComponent {
+export class ClienteDeleteDialogComponent implements OnInit{
     cliente: ICliente;
+    tareas: Tarea[] = [];
+    esPosibleBorrar: boolean;
+    idDelete:string
 
-    constructor(protected clienteService: ClienteService, public activeModal: NgbActiveModal, protected eventManager: JhiEventManager) {}
+    constructor(protected clienteService: ClienteService,
+                public activeModal: NgbActiveModal,
+                protected eventManager: JhiEventManager,
+                protected tareaService: TareaService) {}
 
+    ngOnInit() {
+        this.validarRelacionCliente();
+    }
     clear() {
         this.activeModal.dismiss('cancel');
     }
 
     confirmDelete(id: string) {
-        this.clienteService.delete(id).subscribe(response => {
-            this.eventManager.broadcast({
-                name: 'clienteListModification',
-                content: 'Deleted an cliente'
-            });
-            this.activeModal.dismiss(true);
-        });
+
+      if (this.esPosibleBorrar === true){
+          this.clienteService.delete(id).subscribe(response => {
+              this.eventManager.broadcast({
+                  name: 'clienteListModification',
+                  content: 'Deleted an cliente'
+              });
+              this.activeModal.dismiss(true);
+          });
+      }
+
     }
+    validarRelacionCliente(){
+        this.tareaService.query().subscribe(
+            res=>{
+                this.tareas = res.body;
+                this.restringirbtn();
+            },
+            err=>{
+                console.log(err);
+            }
+        );
+    }
+    restringirbtn(){
+        var i;
+        for (i=0;i<this.tareas.length;i++){
+            if (this.tareas[i].cliente.id === this.cliente.id){
+                this.esPosibleBorrar = false;
+            }
+            else {
+                this.esPosibleBorrar = true;
+            }
+        }
+    }
+
 }
 
 @Component({
@@ -37,10 +76,12 @@ export class ClienteDeleteDialogComponent {
 })
 export class ClienteDeletePopupComponent implements OnInit, OnDestroy {
     protected ngbModalRef: NgbModalRef;
-
-    constructor(protected activatedRoute: ActivatedRoute, protected router: Router, protected modalService: NgbModal) {}
+    constructor(protected activatedRoute: ActivatedRoute,
+                protected router: Router,
+                protected modalService: NgbModal) {}
 
     ngOnInit() {
+
         this.activatedRoute.data.subscribe(({ cliente }) => {
             setTimeout(() => {
                 this.ngbModalRef = this.modalService.open(ClienteDeleteDialogComponent as Component, { size: 'lg', backdrop: 'static' });
@@ -58,8 +99,8 @@ export class ClienteDeletePopupComponent implements OnInit, OnDestroy {
             }, 0);
         });
     }
-
     ngOnDestroy() {
         this.ngbModalRef = null;
     }
+
 }
